BF01 Autenticação (login/cadastro) e chave PIX (BF09) no fron-end e no back-end

+++ SpringPfc/SpringPfc/src/main/java/com/web/pfc/SpringPfc/Repository/InstituicaoRep.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.web.pfc.SpringPfc.domain.Instituicao;
 
 @Repository
-public interface InstituicaoRep extends JpaRepository<Instituicao, Integer> {
+public interface InstituicaoRep extends JpaRepository<Instituicao, Long> {
 
 }
