+++ SpringPfc/SpringPfc/src/main/java/com/web/pfc/SpringPfc/dto/CadUsuarioDTO.java
@@ -1,5 +1,5 @@
 package com.web.pfc.SpringPfc.dto;
 
 public record CadUsuarioDTO(
-        String nome, String email, String senha, Integer instituicaoId, String perfil) {
+        String nome, String email, String senha, Long instituicaoId, String perfil) {
 }
